Replace current onChange
