+++ Sources/Intermodular/Helpers/SwiftUI/View.onChange+.swift
@@ -2,47 +2,47 @@
 // Copyright (c) Vatsal Manot
 //
 
-#if swift(>=5.3)
-
 import Swift
 import SwiftUI
 
-struct OnChangeOfValue<V: Equatable>: ViewModifier {
-    @State var initialValue: V
-    
-    let value: V
-    let action: (V) -> Void
-    
-    func body(content: Content) -> some View {
-        if initialValue != value {
+// Modified from https://stackoverflow.com/questions/58363563/swiftui-get-notified-when-binding-value-changes
+
+struct ChangeObserver<Base: View, Value: Equatable>: View {
+    let base: Base
+    let value: Value
+    @State private var oldValue: Value?
+    let action: (Value?, Value) -> Void
+
+    @State var model = Model()
+
+    var body: some View {
+        if model.update(value: value) {
             DispatchQueue.main.async {
-                initialValue = value
-                
-                action(value)
+                self.action(self.oldValue, self.value)
+                oldValue = value
             }
         }
-        
-        return content
+        return base
+    }
+
+    class Model {
+        private var savedValue: Value?
+        func update(value: Value) -> Bool {
+            guard value != savedValue else {
+                return false
+            }
+            savedValue = value
+            return true
+        }
     }
 }
 
 extension View {
-    @_disfavoredOverload
-    @available(iOS, introduced: 13.0, deprecated: 14.0)
-    @available(macOS, introduced: 10.15, deprecated: 11.0)
-    @available(tvOS, introduced: 13.0, deprecated: 14.0)
-    @available(watchOS, introduced: 6.0, deprecated: 7.0)
-    @ViewBuilder
-    public func onChange<V: Equatable>(
-        of value: V,
-        perform action: @escaping (V) -> Void
-    ) -> some View {
-        if #available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *) {
-            onChange(of: value, perform: action)
-        } else {
-            modifier(OnChangeOfValue(initialValue: value, value: value, action: action))
-        }
+    @available(iOS 13, *)
+    @available(macOS 10.15, *)
+    @available(tvOS 13.0, *)
+    @available(watchOS 6.0, *)
+    func onChange<Value: Equatable>(of value: Value, perform action: @escaping (_ oldValue: Value?, _ newValue: Value) -> Void) -> ChangeObserver<Self, Value> {
+        ChangeObserver(base: self, value: value, action: action)
     }
 }
-
-#endif
